feat: collapse issue cockpit agent spine (PAN-2842)
