+++ src/dashboard/frontend/src/components/Stage/cockpit/agentsLane.module.css
@@ -86,3 +86,29 @@
    info-tone (running/reviewing/testing/verifying) row is visibly alive. */
 .spin { animation: laneSpin 1s linear infinite; vertical-align: -1px; margin-right: 3px; }
 @keyframes laneSpin { to { transform: rotate(360deg); } }
+
+/* The cockpit owns collapse state. Rows remain the same clickable components;
+ * the ancestor attribute only removes their text and centers the icon tiles. */
+@container (min-width: 840px) {
+  :global([data-spine-collapsed="true"]) .header,
+  :global([data-spine-collapsed="true"]) .body,
+  :global([data-spine-collapsed="true"]) .caret,
+  :global([data-spine-collapsed="true"]) .resources {
+    display: none;
+  }
+
+  :global([data-spine-collapsed="true"]) .row {
+    justify-content: center;
+    align-items: center;
+    gap: 0;
+    padding-inline: 0;
+  }
+
+  :global([data-spine-collapsed="true"]) .row.child {
+    margin-left: 0;
+  }
+
+  :global([data-spine-collapsed="true"]) .tile {
+    margin-top: 0;
+  }
+}
